requirements.txt
langchain>=0.0.200
langchain-community>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
llama-cpp-python>=0.2.0
pypdf>=3.9.0
docx2txt>=0.8
flask>=3.0.0
